Remove code duplication and moved some code to JabRefGUI
