+++ jabgui/src/main/java/org/jabref/Launcher.java
@@ -3,38 +3,29 @@ package org.jabref;
 import java.io.File;
 import java.io.IOException;
 import java.net.Authenticator;
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 
 import org.jabref.gui.JabRefGUI;
 import org.jabref.gui.preferences.GuiPreferences;
 import org.jabref.gui.preferences.JabRefGuiPreferences;
-import org.jabref.gui.util.DefaultFileUpdateMonitor;
 import org.jabref.logic.UiCommand;
 import org.jabref.logic.citationstyle.CSLStyleLoader;
-import org.jabref.logic.journals.JournalAbbreviationLoader;
-import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.logic.net.ProxyAuthenticator;
 import org.jabref.logic.net.ProxyPreferences;
 import org.jabref.logic.net.ProxyRegisterer;
 import org.jabref.logic.net.ssl.SSLPreferences;
 import org.jabref.logic.net.ssl.TrustStoreManager;
 import org.jabref.logic.preferences.CliPreferences;
-import org.jabref.logic.protectedterms.ProtectedTermsLoader;
 import org.jabref.logic.remote.RemotePreferences;
 import org.jabref.logic.remote.client.RemoteClient;
 import org.jabref.logic.search.PostgreServer;
 import org.jabref.logic.util.BuildInfo;
 import org.jabref.logic.util.Directories;
-import org.jabref.logic.util.HeadlessExecutorService;
 import org.jabref.migrations.PreferencesMigrations;
-import org.jabref.model.entry.BibEntryTypesManager;
-import org.jabref.model.util.FileUpdateMonitor;
 
 import com.airhacks.afterburner.injection.Injector;
 import org.slf4j.Logger;
@@ -54,7 +45,8 @@ public class Launcher {
     public static void main(String[] args) {
         initLogging(args);
 
-        // Initialize preferences
+        Injector.setModelOrService(BuildInfo.class, new BuildInfo());
+
         final JabRefGuiPreferences preferences = JabRefGuiPreferences.getInstance();
         Injector.setModelOrService(CliPreferences.class, preferences);
         Injector.setModelOrService(GuiPreferences.class, preferences);
@@ -64,23 +56,8 @@ public class Launcher {
             systemExit();
         }
 
-        // ToDo: Not really needed to display available import formats in ArgumentProcessor - Refactor!
-        DefaultFileUpdateMonitor fileUpdateMonitor = new DefaultFileUpdateMonitor();
-        HeadlessExecutorService.INSTANCE.executeInterruptableTask(fileUpdateMonitor, "FileUpdateMonitor");
-        Injector.setModelOrService(FileUpdateMonitor.class, fileUpdateMonitor);
-
-        Injector.setModelOrService(BuildInfo.class, new BuildInfo());
-
-        // ToDo: Needed here? Move to JabRefGUI?
-        BibEntryTypesManager entryTypesManager = preferences.getCustomEntryTypesRepository();
-        Injector.setModelOrService(BibEntryTypesManager.class, entryTypesManager);
-        Injector.setModelOrService(JournalAbbreviationRepository.class, JournalAbbreviationLoader.loadRepository(preferences.getJournalAbbreviationPreferences()));
-        Injector.setModelOrService(ProtectedTermsLoader.class, new ProtectedTermsLoader(preferences.getProtectedTermsPreferences()));
-
         configureProxy(preferences.getProxyPreferences());
         configureSSL(preferences.getSSLPreferences());
-
-        clearOldSearchIndices();
 
         ArgumentProcessor argumentProcessor = new ArgumentProcessor(
                 args,
@@ -90,7 +67,6 @@ public class Launcher {
         // The method `processArguments` quits the whole JVM if no GUI is needed.
         List<UiCommand> uiCommands = argumentProcessor.processArguments();
         if (argumentProcessor.shouldShutDown()) {
-            LOGGER.debug("JabRef shut down after processing command line arguments");
             systemExit();
         }
 
@@ -101,7 +77,7 @@ public class Launcher {
 
         CSLStyleLoader.loadInternalStyles();
 
-        JabRefGUI.setup(uiCommands, preferences, fileUpdateMonitor);
+        JabRefGUI.setup(uiCommands, preferences);
         JabRefGUI.launch(JabRefGUI.class, args);
     }
 
@@ -194,31 +170,5 @@ public class Launcher {
 
     private static void configureSSL(SSLPreferences sslPreferences) {
         TrustStoreManager.createTruststoreFileIfNotExist(Path.of(sslPreferences.getTruststorePath()));
-    }
-
-    private static void clearOldSearchIndices() {
-        Path currentIndexPath = Directories.getFulltextIndexBaseDirectory();
-        Path appData = currentIndexPath.getParent();
-
-        try {
-            Files.createDirectories(currentIndexPath);
-        } catch (IOException e) {
-            LOGGER.error("Could not create index directory {}", appData, e);
-        }
-
-        try (DirectoryStream<Path> stream = Files.newDirectoryStream(appData)) {
-            for (Path path : stream) {
-                if (Files.isDirectory(path) && !path.toString().endsWith("ssl") && path.toString().contains("lucene")
-                        && !path.equals(currentIndexPath)) {
-                    LOGGER.info("Deleting out-of-date fulltext search index at {}.", path);
-                    Files.walk(path)
-                         .sorted(Comparator.reverseOrder())
-                         .map(Path::toFile)
-                         .forEach(File::delete);
-                }
-            }
-        } catch (IOException e) {
-            LOGGER.error("Could not access app-directory at {}", appData, e);
-        }
     }
 }
